Backport old impl
